Ajustes Feitos em Aula

+++ src/main/java/org/example/models/StatusDispositivoModel.java
@@ -5,10 +5,9 @@ import jakarta.persistence.*;
 @Entity
 @Table (name = "StatusDispositivo")
 public class StatusDispositivoModel {
-
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long id;
+    private int id;
 
     private String estado;
 
@@ -22,7 +21,7 @@ public class StatusDispositivoModel {
         return id;
     }
 
-    public void setId(long id) {
+    public void setId(int id) {
         this.id = id;
     }
 
@@ -41,6 +40,4 @@ public class StatusDispositivoModel {
                 ", estado='" + estado + '\'' +
                 '}';
     }
-
-
 }
